log.sh: timestamp update
Now has two global variables configuring date options: one for the file
log, the other for the terminal. Only the format changes, note the
actual timestamp.
The default format for the file is now rfc-3339 up to the nanosecond.
The default format for the tty is '%F %T', e.g. 2019-05-15 17:21:11 (in
local time)