temp.sh: split temp_dir in two functions
This removes problems with using subshells: when calling a function in a
subshell (with command interpolation $() for instance), the function
cannot change global variables (such as the __TEMP_DIR_LIST__ array),
thus making it impossible to track automatically newly created temp
dirs when retrieving the name in a variable (TMP_DIR=$(temp_dir) for
instance).
The temp_dir function now only returns the name of the created temporary
dir. To be tracked and automatically removed, the function
'track_temp_dir' must now be called with the previous temp_dir. Note
that this will remove the directory passed as a parameter without
explicitely checking that it was a temporary directory.